requirements.txt
annotated-types==0.7.0
anyio==4.9.0
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
azure-ai-inference==1.0.0b9
azure-core==1.34.0
blis==1.3.0
Brotli==1.1.0
catalogue==2.0.10
certifi==2025.6.15
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.21.1
colorama==0.4.6
confection==0.1.5
cryptography==45.0.4
cssselect2==0.8.0
cymem==2.0.11
diff-match-patch==20241021
distro==1.9.0
Django==5.2.3
django-filter==25.1
django-import-export==4.3.7
django-widget-tweaks==1.5.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
et_xmlfile==2.0.0
fonttools==4.58.4
gunicorn==23.0.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
idna==3.10
isodate==0.7.2
Jinja2==3.1.6
jiter==0.10.0
langcodes==3.5.0
language_data==1.3.0
lxml==5.4.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
murmurhash==1.0.13
numpy==2.3.1
openai==1.90.0
openpyxl==3.1.5
oscrypto==1.3.0
packaging==25.0
pillow==11.2.1
preshed==3.0.10
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydyf==0.11.0
Pygments==2.19.2
pyHanko==0.29.0
pyhanko-certvalidator==0.27.0
pypdf==5.6.0
pyphen==0.17.2
python-bidi==0.6.6
python-telegram-bot==22.1
PyYAML==6.0.2
RapidFuzz==3.13.0
reportlab==4.4.1
requests==2.32.4
rich==14.0.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.3
srsly==2.5.1
svglib==1.5.1
tablib==3.8.0
thefuzz==0.22.1
thinc==8.3.6
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.4.0
wasabi==1.1.3
weasel==0.4.1
weasyprint==65.1
webencodings==0.5.1
whitenoise==6.9.0
wrapt==1.17.2
xhtml2pdf==0.2.17
zopfli==0.2.3.post1
